+++ connect/connect-sftp-sink/sftp-sink-ssh-pem-file.sh
@@ -4,6 +4,13 @@ set -e
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 source ${DIR}/../../scripts/utils.sh
 
+if version_gt $TAG_BASE "7.9.99" && ! version_gt $CONNECTOR_TAG "3.1.99"
+then
+     logwarn "minimal supported connector version is 3.2.0 for CP 8.0"
+     logwarn "see https://docs.confluent.io/platform/current/connect/supported-connector-version-8.0.html#supported-connector-versions-in-cp-8-0"
+     exit 111
+fi
+
 log "generate ssh host key"
 rm -rf ssh_host*
 ssh-keygen -t rsa -b 4096 -f ssh_host_rsa_key -P "mypassword"
